Move format relative date to utils
